setup spatie authorization + initial donation category

+++ app/Http/Controllers/DonationCategoryController.php
@@ -2,32 +2,43 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\DonationCategory;
 use Illuminate\Http\Request;
 
 class DonationCategoryController extends Controller
 {
     public function index(Request $request)
     {
-        $data = $request->user()->donation_categories;
+        $data = DonationCategory::generateTree();
         return $data;
     }
 
     public function add(Request $request)
     {
-        $data = $request->user()->donation_categories()->create($request->validated());
+        $data = DonationCategory::create($request->all());
         return $data;
     }
 
     public function edit(Request $request, $id)
     {
-        $data = $request->user()->donation_categories->find($id);
-        $data->update($request->validated());
+        $data = DonationCategory::find($id);
+        if (!$data) {
+            return response()->json([
+                'message' => 'Category not found',
+            ])->setStatusCode(404);
+        }
+        $data->update($request->all());
         return $data;
     }
 
     public function delete(Request $request, $id)
     {
-        $data = $request->user()->donation_categories->find($id);
+        $data = DonationCategory::find($id);
+        if (!$data) {
+            return response()->json([
+                'message' => 'Category not found',
+            ])->setStatusCode(404);
+        }
         $data->delete();
         return response()->json([
             'message' => 'Category deleted successfully',
